requirements.txt
functions-framework==3.*
google-api-python-client==2.*
python-telegram-bot==20.*
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.*
google-generativeai==0.3.*
youtube-transcript-api==0.6.*